Detect if defaults supplied accept header.
A Restful API I use only likes one Accept header value or it will default to XML.  This change detects if the user supplied an Accept header, it will then just clear the default and not add the ones specified in the factory.

Note, I'm on an airplane and my Visual Studio is acting flakey on my laptop, so I couldn't run unit tests or even compile.. don't ask.. have old laptop and I don't have latest update of Studio 2013 or 2015 and I couldn't open the main project file.   And to top it off this is my first commit to git.. ever..  Great project, going to save me a ton of time.